found bug for 2 node system.  locally elected leaders are declared for next phase of elections by when a response message is received from left and right side.   In a 2 node system, we cannot determine if a message is from left or right.   We could count number of responses but in the 2 node case we will just assume that any single response and the fact that it is a two node system means the receipient is a locally elected leader.
Also note that initial value of phase was 0 and distance was pow(2,0) =1.
outgoing messages were only allowed to decrement to 1 so no message
was allowed to travel further than initial node and would decrment to 0.
changed checks to <=0 for message distance from ==1.  Also initialed phase
to 1 so that initial allowable distance is pow(2,1)